+++ Arduino/Layering/ThirdPartyIncludes/02-serial_setup/serial_com.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <Arduino.h>
 
-void setupSerial(int baud_rate);
+void setupSerial(unsigned long baud_rate);
